// 23.3/test210500/bundle/src/main/webapp/libs/com-example-test210500/src/lib/view-components/test-debug-component/design/test-debug-component.interface.ts
export interface ITestDebugComponentParameters {
  viewName: string;
  recordName: string;
  processName: string;
  namedListName: string;
  associationName: string;
  chatbotName: string;
  textValue: string;
  numberValue: string;
  booleanValue: boolean;
  expressionValue: string;
  textAreaValue: string;
  switchValue: boolean;
  selectValue: string;
  colorPickerValue: string;
  tagsValue: string;
  gradient: string;
}